test: split to avoid test failed
